requirements.txt
get==1.0.2
post==1.0.1
public==1.0.2
query-string==1.0.1
request==1.0.1
